backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
Alembic==1.12.0
PyMySQL==1.1.0
cryptography==42.0.5
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6
redis==5.0.0
WeasyPrint==60.0
openpyxl==3.1.2
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Pillow==10.0.0
boto3==1.28.62
flower==2.0.1
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0
email-validator==2.1.1
pandas==2.2.2